Moved LazyImage into common/components folder (also did for Campaigns repository). Fixed an updateState bug in LazyImage.
